+++ elanmanager/impl/src/main/java/org/opendaylight/netvirt/elan/utils/ElanUtils.java
@@ -313,7 +313,7 @@ public class ElanUtils {
             Future<RpcResult<AllocateIdOutput>> result = idManager.allocateId(getIdInput);
             RpcResult<AllocateIdOutput> rpcResult = result.get();
             if (rpcResult.isSuccessful()) {
-                return Uint32.valueOf(rpcResult.getResult().getIdValue().longValue());
+                return rpcResult.getResult().getIdValue();
             } else {
                 LOG.warn("RPC Call to Allocate Id returned with Errors {}", rpcResult.getErrors());
             }
